requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
docopt==0.6.2
fastapi==0.88.0
h11==0.14.0
httptools==0.5.0
idna==3.4
numpy==1.23.5
Pillow==9.3.0
pipreqs==0.4.11
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarg==0.1.9
